Fix failing of test in account page
