+++ tests/e2e/specs/sendPage.js
@@ -118,6 +118,7 @@ describe("Test cases for Send Page", () => {
         .should('contain','Enter amount')
         .get('.ae-modal-light button')
         .click()
+        .wait(5000)
         .get('.ae-input.aemount')
         .type('0.00000001')
         .get('.sendBtn')
@@ -144,6 +145,7 @@ describe("Test cases for Send Page", () => {
         .click()
         .get('.ae-address-input')
         .type(PUBLIC_KEY_SEND)
+        .wait(5000)
         .get('.ae-input.aemount')
         .type('0.00000001')
         .get('.sendBtn')
